requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
Django==4.2
django-extensions==3.2.1
djangorestframework==3.14.0
flake8==6.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
